Simplify build system and add Property Value Converters for Umbraco 7
