--- BuildPackage/package.proj
@@ -1,182 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003" DefaultTargets="Package">
-
-	<!-- IMPORTS -->
-	<PropertyGroup>
-		<MSBuildCommunityTasksPath>$(MSBuildProjectDirectory)\Tools\MSBuildCommunityTasks</MSBuildCommunityTasksPath>
-		<MSBuildUmbracoTasksPath>$(MSBuildProjectDirectory)\Tools\MSBuildUmbracoTasks</MSBuildUmbracoTasksPath>
-		<MSBuildNugetTasksPath>$(MSBuildProjectDirectory)\Tools\MSBuildNugetTasks</MSBuildNugetTasksPath>
-		<AppVeyorUmbracoPath>$(MSBuildProjectDirectory)\Tools\AppVeyorUmbraco</AppVeyorUmbracoPath>
-	</PropertyGroup>
-
-  <Import Project="$(MSBuildCommunityTasksPath)\MSBuild.Community.Tasks.Targets" />
-	<Import Project="$(MSBuildUmbracoTasksPath)\MSBuild.Umbraco.Tasks.Targets" />
-	<Import Project="$(MSBuildNugetTasksPath)\MSBuild.NuGet.Tasks.Targets" />
-	<Import Project="$(AppVeyorUmbracoPath)\AppVeyorUmbraco.Targets" />
-
-	<!-- SHARED PROPERTIES -->
-	<PropertyGroup>
-		<PackageName>Nesting Contently</PackageName>
-		<MinUmbracoVersion>7.7.0</MinUmbracoVersion>
-		<Readme>A property editor for toggling the display state of Nested Content items</Readme>
-		<AuthorName>Nathan Woulfe</AuthorName>
-		<AuthorUrl>https://github.com/nathanwoulfe</AuthorUrl>
-		<PackageLicenseName>MIT license</PackageLicenseName>
-		<PackageLicenseUrl>http://www.opensource.org/licenses/mit-license.php</PackageLicenseUrl>
-		<ProjectUrl>https://github.com/nathanwoulfe/nestingcontently</ProjectUrl>
-	</PropertyGroup>
-
-	<!-- NUGET ONLY PROPERTIES -->
-	<PropertyGroup>
-		<NuGetPackageName>Nesting Contently</NuGetPackageName>
-		<PackageId>NestingContently.Umbraco</PackageId>
-		<Copyright>Nathan Woulfe</Copyright>
-		<Owners>Nathan Woulfe</Owners>
-		<Tags>Umbraco UmbracoCMS NestedContent</Tags>
-		<IconUrl>http://icon.com/img.png</IconUrl>
-		<Language>en-GB</Language>
-		<RequireLicenseAcceptance>false</RequireLicenseAcceptance>
-	</PropertyGroup>
-
-  <!-- APPVEYOR PROPERTIES -->
-  <PropertyGroup>
-    <FileVersion>$(APPVEYOR_BUILD_VERSION)</FileVersion>
-  </PropertyGroup>
-  <Choose>
-    <When Condition="$(APPVEYOR_BUILD_NUMBER) != '' And $(APPVEYOR_REPO_TAG) != 'true' ">
-      <PropertyGroup>
-        <Release>false</Release>
-      </PropertyGroup>
-    </When>
-    <Otherwise>
-      <PropertyGroup>
-        <Release>true</Release>
-      </PropertyGroup>
-    </Otherwise>
-  </Choose>
-
-  <Choose>
-    <When Condition="$(Release) == 'false' And $(UMBRACO_PACKAGE_PRERELEASE_SUFFIX) == 'rtm'">
-      <PropertyGroup>
-        <AbortBuild>true</AbortBuild>
-      </PropertyGroup>
-    </When>
-    <Otherwise>
-      <PropertyGroup>
-        <AbortBuild>false</AbortBuild>
-      </PropertyGroup>
-    </Otherwise>
-  </Choose> 
-
-  <!-- PATHS -->
-  <PropertyGroup>
-    <BuildConfig>Release</BuildConfig>
-    <RootDir>$(MSBuildProjectDirectory)\..</RootDir>
-    <BuildUmbDir>$(MSBuildProjectDirectory)\_umbraco</BuildUmbDir>
-    <BuildNuGetDir>$(MSBuildProjectDirectory)\_nuget</BuildNuGetDir>
-    <ArtifactsDir>$(MSBuildProjectDirectory)\artifacts</ArtifactsDir>
-    <GruntProjectDir>$(RootDir)\dist</GruntProjectDir>
-  </PropertyGroup>
-
-  <!-- TARGETS -->
-  <Target Name="GetProductVersion">
-    <Error Condition="$(AbortBuild) == 'true'" Text="Aborting the build as the UMBRACO_PACKAGE_PRERELEASE_SUFFIX suffix is set 'rtm' but APPVEYOR_REPO_TAG is not 'true'" />
-    <GetProductVersion BuildVersion="$(APPVEYOR_BUILD_VERSION)" BuildSuffix="$(UMBRACO_PACKAGE_PRERELEASE_SUFFIX)" Release="$(Release)">
-      <Output TaskParameter="ProductVersion" PropertyName="ProductVersion"/>
-    </GetProductVersion>
-  </Target>
-
-  <!-- CLEAN -->
-  <Target Name="Clean" DependsOnTargets="GetProductVersion">
-    <RemoveDir Directories="$(BuildUmbDir)" Condition="Exists('$(BuildUmbDir)')" />
-    <RemoveDir Directories="$(BuildNuGetDir)" Condition="Exists('$(BuildNuGetDir)')" />
-    <RemoveDir Directories="$(ArtifactsDir)" Condition="Exists('$(ArtifactsDir)')" />
-    <MakeDir Directories="$(BuildUmbDir)" />
-    <MakeDir Directories="$(BuildNuGetDir)" />
-    <MakeDir Directories="$(ArtifactsDir)" />
-  </Target>
-
-
-	<!-- PREPARE FILES -->
-	<Target Name="PrepareFiles" DependsOnTargets="GetProductVersion">
-		<ItemGroup>
-			<PackageFile Include="$(MSBuildProjectDirectory)\package.xml" />
-			<NuSpecFile Include="$(MSBuildProjectDirectory)\package.nuspec" />
-			<GruntProjectFiles Include="$(GruntProjectDir)\**\*" />
-		</ItemGroup>
-		<Copy SourceFiles="@(PackageFile)" DestinationFolder="$(BuildUmbDir)" />
-		<Copy SourceFiles="@(NuSpecFile)" DestinationFolder="$(BuildNuGetDir)" />
-		<Copy SourceFiles="@(GruntProjectFiles)" DestinationFiles="@(GruntProjectFiles->'$(BuildUmbDir)\%(RecursiveDir)%(Filename)%(Extension)')" />
-		<Copy SourceFiles="@(GruntProjectFiles)" DestinationFiles="@(GruntProjectFiles->'$(BuildNuGetDir)\Content\%(RecursiveDir)%(Filename)%(Extension)')" />    
-	  </Target>
-
-  <!-- MANIFEST UMBRACO -->
-  <Target Name="ManifestUmbraco" DependsOnTargets="PrepareFiles">
-    <ItemGroup>
-      <ManifestFiles Include="$(BuildUmbDir)\**\*" Exclude="$(BuildUmbDir)\package.xml" />
-    </ItemGroup>
-    <ManifestUpdate
-			ManifestFile="$(BuildUmbDir)\package.xml"
-			WorkingDirectory="$(BuildUmbDir)"
-			MinimumRequiredUmbracoVersion="$(MinUmbracoVersion)"
-			PackageName="$(PackageName)"
-			PackageVersion="$(ProductVersion)"
-			AuthorName="$(AuthorName)"
-			AuthorUrl="$(AuthorUrl)"
-			Readme="$(Readme)"
-			PackageLicenseName="$(PackageLicenseName)"
-			PackageLicenseUrl="$(PackageLicenseUrl)"
-			PackageUrl="$(ProjectUrl)"
-			Files="@(ManifestFiles)" />
-  </Target>
-
-  <!-- MANIFEST FOR NUGET PACKAGE -->
-  <Target Name="ManifestNuGet" DependsOnTargets="PrepareFiles">
-    <ItemGroup>
-      <ManifestFiles Include="$(BuildNuGetDir)\**\*" Exclude="$(BuildNuGetDir)\package.nuspec" />
-    </ItemGroup>
-    <MSBuild.NuGet.Tasks.ManifestUpdate
-			ManifestFile="$(BuildNuGetDir)\package.nuspec"
-			WorkingDirectory="$(BuildNuGetDir)"
-			Title="$(PackageName)"
-			Description="$(Readme)"
-			Summary="$(Readme)"
-			Version="$(ProductVersion)"
-			MinimumRequiredUmbracoVersion="$(MinUmbracoVersion)"
-			Authors="$(AuthorName)"
-			Owners="$(Owners)"
-			Copyright="$(Copyright)"
-			LicenseUrl="$(PackageLicenseUrl)"
-			ProjectUrl="$(ProjectUrl)"
-			Id="$(PackageId)"
-			IconUrl="$(IconUrl)"
-			Language="$(Language)"
-			RequireLicenseAcceptance="$(RequireLicenseAcceptance)"
-			Tags="$(Tags)"
-			Files="@(ManifestFiles)" />
-  </Target>
-
-  <!-- PACKAGE -->
-  <Target Name="Package" DependsOnTargets="ManifestUmbraco; ManifestNuGet">
-    <ItemGroup>
-      <PackageFiles Include="$(BuildUmbDir)\**\*.*" />
-    </ItemGroup>
-
-    <Package ManifestFile="$(BuildUmbDir)\package.xml"
-			WorkingDirectory="$(BuildUmbDir)"
-			OutputDirectory="$(ArtifactsDir)"
-			Files="@(PackageFiles)" />
-
-    <MSBuild.NuGet.Tasks.Pack NuGetExePath="$(RootDir)\BuildPackage\Tools\NuGet.exe"
-          ManifestFile="$(BuildNuGetDir)\package.nuspec"
-          BasePath="$(BuildNuGetDir)"
-          Version="$(ProductVersion)"
-          OutputDirectory="$(ArtifactsDir)"
-          Symbols="true" />
-
-    <RemoveDir Directories="$(BuildUmbDir)" Condition="Exists('$(BuildUmbDir)')" />
-    <RemoveDir Directories="$(BuildNuGetDir)" Condition="Exists('$(BuildNuGetDir)')" />
-
-  </Target>
-
-</Project>
