+++ NestingContently.Umbraco/App_Plugins/NestingContently/backoffice/editor.html
@@ -1,0 +1,3 @@
+<div ng-controller="nestingContentlyController as vm">
+    {{ model.value }}
+</div>
