+++ NestingContently.Umbraco/src/entry.js
@@ -1,0 +1,3 @@
+import "./editor.controller";
+import "./interceptor";
+import "./button.directive";
