--- src/NestingContently/backoffice/editor.html
@@ -1,3 +1,0 @@
-<div ng-controller="nestingContentlyController as vm">
-    {{ model.value }}
-</div>
