--- src/NestingContently/package.dev.manifest
@@ -1,21 +1,0 @@
-﻿{   
-    propertyEditors: [      
-        {
-            alias: "NestingContently",
-            name: "Nesting Contently",
-            editor: {
-                view: "~/app_plugins/nestingcontently/backoffice/editor.html",
-				        valueType: "INT"
-            }
-        }        
-    ],
-    css: [
-        '~/app_plugins/nestingcontently/backoffice/styles.css'
-    ],
-    javascript: [	
-        '~/app_plugins/nestingcontently/backoffice/editor.controller.js',
-        '~/app_plugins/nestingcontently/backoffice/interceptor.js',
-        '~/app_plugins/nestingcontently/backoffice/button.directive.js'
-
-    ]    
-}
