--- src/NestingContently/package.prod.manifest
@@ -1,18 +1,0 @@
-﻿{  
-    propertyEditors: [      
-        {
-            alias: "NestingContently",
-            name: "Nesting Contently",
-            editor: {
-                view: "~/app_plugins/nestingcontently/backoffice/editor.html",
-				        valueType: "INT"
-            }
-        }        
-    ],
-    css: [
-        '~/app_plugins/nestingcontently/backoffice/styles.css'
-    ],
-    javascript: [	
-		      '~/app_plugins/nestingcontently/backoffice/nesting-contently.min.js'
-    ]
-}
